Move Commitments service to shared
