+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.Commitments.Shared/Models/CohortDetails.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ProviderCommitments.Models
+﻿namespace SFA.DAS.Commitments.Shared.Models
 {
     public class CohortDetails
     {
