+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.Commitments.Shared/Models/EditDraftApprenticeshipDetails.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SFA.DAS.ProviderCommitments.Models.ApiModels
+namespace SFA.DAS.Commitments.Shared.Models
 {
     public class EditDraftApprenticeshipDetails
     {
