+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.Commitments.Shared/SFA.DAS.Commitments.Shared.csproj
@@ -10,6 +10,7 @@
     <PackageReference Include="Moq" Version="4.10.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
     <PackageReference Include="SFA.DAS.CommitmentsV2.Api.Client" Version="2.1.1462" />
+    <PackageReference Include="SFA.DAS.Encoding" Version="1.1.1" />
   </ItemGroup>
 
 </Project>
