+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.UnitTests/Services/ProviderCommitmentsServiceTests/ProviderCommitmentsServiceTestFixtures.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading;
 using AutoFixture;
 using Moq;
+using SFA.DAS.Commitments.Shared.Services;
 using SFA.DAS.CommitmentsV2.Api.Client;
 using SFA.DAS.CommitmentsV2.Api.Types.Requests;
 using SFA.DAS.CommitmentsV2.Api.Types.Responses;
@@ -22,7 +23,7 @@ namespace SFA.DAS.ProviderCommitments.UnitTests.Services.ProviderCommitmentsServ
             AddDraftApprenticeshipRequest = autoFixture.Build<AddDraftApprenticeshipRequest>().Create();
             GetDraftApprenticeshipResponse = autoFixture.Build<GetDraftApprenticeshipResponse>().Create();
 
-            Sut = new ProviderCommitmentsService(CommitmentsApiClientMock.Object, HashingServiceMock.Object);
+            Sut = new CommitmentsService(CommitmentsApiClientMock.Object, HashingServiceMock.Object);
         }
 
         public long CohortId { get; }
@@ -32,7 +33,7 @@ namespace SFA.DAS.ProviderCommitments.UnitTests.Services.ProviderCommitmentsServ
         public Mock<ICommitmentsApiClient> CommitmentsApiClientMock { get; }
         public Mock<IEncodingService> HashingServiceMock{ get; }
 
-        public ProviderCommitmentsService Sut;
+        public CommitmentsService Sut;
 
         public ProviderCommitmentsServiceTestFixtures SetupGetCohortDetailsReturnValue(GetCohortResponse retVal)
         {
