+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web.UnitTests/Mappers/EditDraftApprenticeshipDetailsToViewModelMapperTests/WhenIMapDraftApprenticeshipToCohortRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using AutoFixture;
 using NUnit.Framework;
-using SFA.DAS.ProviderCommitments.Models.ApiModels;
+using SFA.DAS.Commitments.Shared.Models;
 using SFA.DAS.ProviderCommitments.Web.Mappers;
 using SFA.DAS.ProviderCommitments.Web.Models;
