+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Mappers/EditDraftApprenticeshipDetailsToViewModelMapper.cs
@@ -1,5 +1,5 @@
 ﻿using SFA.DAS.Commitments.Shared.Interfaces;
-using SFA.DAS.ProviderCommitments.Models.ApiModels;
+using SFA.DAS.Commitments.Shared.Models;
 using SFA.DAS.ProviderCommitments.Web.Models;
 
 namespace SFA.DAS.ProviderCommitments.Web.Mappers
