+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Validators/MonthYearValidator.cs
@@ -1,6 +1,5 @@
 ﻿using FluentValidation;
 using SFA.DAS.Commitments.Shared.Models;
-using SFA.DAS.ProviderCommitments.Models;
 
 namespace SFA.DAS.ProviderCommitments.Web.Validators
 {
