+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.csproj
@@ -11,6 +11,7 @@
 
   <ItemGroup>
     <Folder Include="Application\" />
+    <Folder Include="Models\" />
   </ItemGroup>
 
   <ItemGroup>
